Fix browser crash for MSM7x27
Flag ENABLE_V8_SNAPSHOT should not be enabled for MSM7x27 target.
TARGET_ARCH_VARIANT is used to identify the targets that should not have
the flag enabled in the Android v8 top-level makefile.  Recently,
TARGET_ARCH_VARIANT has been set to armv6k-vfp for MSM7x27.  Change the
check in the makefile to reflect this.

Change-Id: I8f7ea71b9b532d02d882c531810acb84528b5d54