v8: prevent unnecessary rebuilds.
Android.mksnapshot.mk is parsed twice when building for a 64-bit device,
for example once for arm64 and again for arm.  Using += for
LOCAL_CFLAGS_v8_target_* results in cflags being added twice for the
32-bit version.  For a 32-bit device the file is only parsed once and
the cflags are not duplicated.

The cflags are used for host modules, which are shared between builds of
different products.  If the cflags change between products the command
line will change and ninja will rebuild all the files.

Replace += with := for the first assignment to LOCAL_CFLAGS_v8_target_*.

Bug: 26470151
Bug: 31987131
Bug: 29178923

Change-Id: Idb7bc157830207fe5314b5b084a0bd564d5f3221
(cherry picked from commit 665246c65f856a1a352bd4fd8e843ce4b4c72d3d)